requirements.txt
pdf2image
fastapi 
uvicorn
pytesseract 
Pillow
opencv-python
numpy